review-service/pyproject.toml
[tool.poetry]
name = "book-service"
version = "0.1.0"
description = ""
authors = ["Maks1mka <maksim_goncharov_2006@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.119.0"
uvicorn = "^0.38.0"
sqlalchemy = "==2.0.44"
asyncpg = "==0.30.0"
python-dotenv = "==1.1.1"
pydantic = "^2.0.0"
PyPDF2 = "^3.0.0"
python-multipart = "^0.0.6"
aiofiles = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
